src/app/components/services/todo.service.ts
import { Injectable } from '@angular/core';

@Injectable({
  providedIn: 'root'
})
export class TodoService {

  constructor(
    // // Bind `fetch` to `window` to ensure that `window` is the `this` context
    // private fetch = window.fetch.bind(window)
  ) { }

  // public async getTodos(): Promise<string[]> {
  //   const response = await this.fetch('/todos');
  //   if (!response.ok) {
  //     throw new Error(
  //       `HTTP error: ${response.status} ${response.statusText}`
  //     );
  //   }
  //   return await response.json();
  // }

    public async getTodos(): Promise<string[]> {
      const response = await fetch('/todos');
      if (!response.ok) {
        throw new Error(
          `HTTP error: ${response.status} ${response.statusText}`
        );
      }
      return await response.json();
    }
}
